+++ src/xclim/indicators/atmos/_precip.py
@@ -4,7 +4,7 @@ from __future__ import annotations
 
 from xarray import DataArray
 
-from xclim import indices
+from xclim import compute
 from xclim.core import cfchecks
 from xclim.core.indicator import (
     Daily,
@@ -142,7 +142,7 @@ rain_on_frozen_ground_days = PrTasxWithIndexing(
     abstract="The number of days with rain above a given threshold after a series of seven days with average daily "
     "temperature below 0°C. Precipitation is assumed to be rain when the daily average temperature is above 0°C.",
     cell_methods="",
-    compute=indices.rain_on_frozen_ground_days,
+    compute=compute.rain_on_frozen_ground_days,
 )
 
 max_1day_precipitation_amount = PrecipWithIndexing(
@@ -154,7 +154,7 @@ max_1day_precipitation_amount = PrecipWithIndexing(
     description="{freq} maximum 1-day total precipitation",
     abstract="Maximum total daily precipitation for a given period.",
     cell_methods="time: maximum over days",
-    compute=indices.max_1day_precipitation_amount,
+    compute=compute.max_1day_precipitation_amount,
 )
 
 max_n_day_precipitation_amount = Precip(
@@ -167,7 +167,7 @@ max_n_day_precipitation_amount = Precip(
     description="{freq} maximum {window}-day total precipitation amount.",
     abstract="Maximum of the moving sum of daily precipitation for a given period.",
     cell_methods="time: maximum over days",
-    compute=indices.max_n_day_precipitation_amount,
+    compute=compute.max_n_day_precipitation_amount,
 )
 
 wetdays = PrecipWithIndexing(
@@ -179,7 +179,7 @@ wetdays = PrecipWithIndexing(
     description="{freq} number of days with daily precipitation at or above {thresh}.",
     abstract="The number of days with daily precipitation at or above a given threshold.",
     cell_methods="time: sum over days",
-    compute=indices.wetdays,
+    compute=compute.wetdays,
 )
 
 wetdays_prop = PrecipWithIndexing(
@@ -190,7 +190,7 @@ wetdays_prop = PrecipWithIndexing(
     description="{freq} proportion of days with precipitation at or above {thresh}.",
     abstract="The proportion of days with daily precipitation at or above a given threshold.",
     cell_methods="time: sum over days",
-    compute=indices.wetdays_prop,
+    compute=compute.wetdays_prop,
 )
 
 dry_days = PrecipWithIndexing(
@@ -202,7 +202,7 @@ dry_days = PrecipWithIndexing(
     description="{freq} number of days with daily precipitation under {thresh}.",
     abstract="The number of days with daily precipitation under a given threshold.",
     cell_methods="time: sum over days",
-    compute=indices.dry_days,
+    compute=compute.dry_days,
 )
 
 dryness_index = Precip(
@@ -218,7 +218,7 @@ dryness_index = Precip(
     "Metric originally published in Riou et al. (1994).",
     cell_methods="",
     src_freq=["D", "MS"],
-    compute=indices.dryness_index,
+    compute=compute.dryness_index,
 )
 
 maximum_consecutive_wet_days = Precip(
@@ -230,7 +230,7 @@ maximum_consecutive_wet_days = Precip(
     description="{freq} maximum number of consecutive days with daily precipitation at or above {thresh}.",
     abstract="The longest number of consecutive days where daily precipitation is at or above a given threshold.",
     cell_methods="time: sum over days",
-    compute=indices.maximum_consecutive_wet_days,
+    compute=compute.maximum_consecutive_wet_days,
 )
 
 maximum_consecutive_dry_days = Precip(
@@ -242,7 +242,7 @@ maximum_consecutive_dry_days = Precip(
     description="{freq} maximum number of consecutive days with daily precipitation below {thresh}.",
     abstract="The longest number of consecutive days where daily precipitation below a given threshold.",
     cell_methods="time: sum over days",
-    compute=indices.maximum_consecutive_dry_days,
+    compute=compute.maximum_consecutive_dry_days,
 )
 
 daily_pr_intensity = PrecipWithIndexing(
@@ -256,7 +256,7 @@ daily_pr_intensity = PrecipWithIndexing(
     "{freq} average precipitation for days with daily precipitation over {thresh}.",
     abstract="Average precipitation for days with daily precipitation above a given threshold.",
     cell_methods="",
-    compute=indices.daily_pr_intensity,
+    compute=compute.daily_pr_intensity,
 )
 
 max_pr_intensity = HrPrecip(
@@ -268,7 +268,7 @@ max_pr_intensity = HrPrecip(
     description="{freq} maximum precipitation intensity over rolling {window}h time window.",
     abstract="Maximum precipitation intensity over a given rolling time window.",
     cell_methods="time: max",
-    compute=indices.max_pr_intensity,
+    compute=compute.max_pr_intensity,
     duration="{window}",
     keywords="IDF curves",
 )
@@ -284,7 +284,7 @@ precip_accumulation = PrecipWithIndexing(
     "used to restrict the calculation to precipitation of only one phase (liquid or solid). Precipitation is "
     "considered solid if the average daily temperature is below 0°C (and vice versa).",
     cell_methods="time: sum over days",
-    compute=indices.precip_accumulation,
+    compute=compute.precip_accumulation,
     parameters={"tas": None, "phase": None, "thresh": None},
 )
 
@@ -299,7 +299,7 @@ precip_average = PrecipWithIndexing(
     "used to restrict the calculation to precipitation of only one phase (liquid or solid). Precipitation is "
     "considered solid if the average daily temperature is below 0°C threshold (and vice versa).",
     cell_methods="time: mean over days",
-    compute=indices.precip_average,
+    compute=compute.precip_average,
     parameters={"tas": None, "phase": None},
 )
 
@@ -313,7 +313,7 @@ wet_precip_accumulation = PrecipWithIndexing(
     abstract="Total accumulated precipitation on days with precipitation. "
     "A day is considered to have precipitation if the precipitation is greater than or equal to a given threshold.",
     cell_methods="time: sum over days",
-    compute=indices.prcptot,
+    compute=compute.prcptot,
     parameters={"thresh": {"default": "1 mm/day"}},
 )
 
@@ -327,7 +327,7 @@ liquid_precip_accumulation = PrTasxWithIndexing(
     abstract="Total accumulated liquid precipitation. "
     "Precipitation is considered liquid when the average daily temperature is above a given threshold.",
     cell_methods="time: sum over days",
-    compute=indices.precip_accumulation,
+    compute=compute.precip_accumulation,
     parameters={"tas": {"kind": InputKind.VARIABLE}, "phase": "liquid"},
 )
 
@@ -341,7 +341,7 @@ liquid_precip_average = PrTasxWithIndexing(
     abstract="Averaged liquid precipitation. "
     "Precipitation is considered liquid when the average daily temperature is above a given threshold.",
     cell_methods="time: mean over days",
-    compute=indices.precip_average,
+    compute=compute.precip_average,
     parameters={"tas": {"kind": InputKind.VARIABLE}, "phase": "liquid"},
 )
 
@@ -355,7 +355,7 @@ solid_precip_accumulation = PrTasxWithIndexing(
     abstract="Total accumulated solid precipitation. "
     "Precipitation is considered solid when the average daily temperature is at or below a given threshold.",
     cell_methods="time: sum over days",
-    compute=indices.precip_accumulation,
+    compute=compute.precip_accumulation,
     parameters={"tas": {"kind": InputKind.VARIABLE}, "phase": "solid"},
 )
 
@@ -369,7 +369,7 @@ solid_precip_average = PrTasxWithIndexing(
     abstract="Averaged solid precipitation. "
     "Precipitation is considered solid when the average daily temperature is at or below a given threshold.",
     cell_methods="time: mean over days",
-    compute=indices.precip_average,
+    compute=compute.precip_average,
     parameters={"tas": {"kind": InputKind.VARIABLE}, "phase": "solid"},
 )
 
@@ -386,7 +386,7 @@ standardized_precipitation_index = StandardizedIndexes(
     "The window unit `X` is the minimal time period defined by the resampling frequency.",
     cell_methods="",
     keywords="precipitation",
-    compute=indices.standardized_precipitation_index,
+    compute=compute.standardized_precipitation_index,
 )
 
 standardized_precipitation_evapotranspiration_index = StandardizedIndexes(
@@ -404,7 +404,7 @@ standardized_precipitation_evapotranspiration_index = StandardizedIndexes(
     "resampling frequency.",
     cell_methods="",
     keywords="precipitation",
-    compute=indices.standardized_precipitation_evapotranspiration_index,
+    compute=compute.standardized_precipitation_evapotranspiration_index,
 )
 
 drought_code = FireWeather(
@@ -415,7 +415,7 @@ drought_code = FireWeather(
     description="Numerical code estimating the average moisture content of organic layers.",
     abstract="The Drought Index is part of the Canadian Forest-Weather Index system. "
     "It is a numerical code that estimates the average moisture content of organic layers.",
-    compute=indices.drought_code,
+    compute=compute.drought_code,
     missing="skip",
 )
 
@@ -424,7 +424,7 @@ duff_moisture_code = FireWeather(
     units="",
     long_name="Duff Moisture Code",
     description="Numeric rating of the average moisture content of loosely compacted organic layers of moderate depth.",
-    compute=indices.duff_moisture_code,
+    compute=compute.duff_moisture_code,
     missing="skip",
 )
 
@@ -457,7 +457,7 @@ cffwis_indices = FireWeather(
         "Numeric rating of fire intensity.",
     ],
     units="",
-    compute=indices.cffwis_indices,
+    compute=compute.cffwis_indices,
     missing="skip",
 )
 
@@ -468,7 +468,7 @@ keetch_byram_drought_index = FireWeather(
     standard_name="keetch_byram_drought_index",
     long_name="Keetch-Byran Drought Index",
     description="Amount of water necessary to bring the soil moisture content back to field capacity",
-    compute=indices.keetch_byram_drought_index,
+    compute=compute.keetch_byram_drought_index,
     missing="skip",
 )
 
@@ -479,7 +479,7 @@ griffiths_drought_factor = FireWeather(
     standard_name="griffiths_drought_factor",
     long_name="Griffiths Drought Factor",
     description="Numeric indicator of the forest fire fuel availability in the deep litter bed",
-    compute=indices.griffiths_drought_factor,
+    compute=compute.griffiths_drought_factor,
     missing="skip",
 )
 
@@ -490,7 +490,7 @@ mcarthur_forest_fire_danger_index = FireWeather(
     standard_name="mcarthur_forest_fire_danger_index",
     long_name="McArthur Forest Fire Danger Index",
     description="Numeric rating of the potential danger of a forest fire",
-    compute=indices.mcarthur_forest_fire_danger_index,
+    compute=compute.mcarthur_forest_fire_danger_index,
     missing="skip",
 )
 
@@ -504,7 +504,7 @@ last_snowfall = PrecipWithIndexing(
     abstract="The last day where snowfall exceeded a given threshold during a time period (the threshold can be "
     "given as a snowfall flux or a liquid water equivalent snowfall rate).",
     units="",
-    compute=indices.last_snowfall,
+    compute=compute.last_snowfall,
 )
 
 first_snowfall = PrecipWithIndexing(
@@ -516,7 +516,7 @@ first_snowfall = PrecipWithIndexing(
     abstract="The first day where snowfall exceeded a given threshold during a time period (the threshold can be "
     "given as a snowfall flux or a liquid water equivalent snowfall rate).",
     units="",
-    compute=indices.first_snowfall,
+    compute=compute.first_snowfall,
 )
 
 days_with_snow = PrecipWithIndexing(
@@ -526,7 +526,7 @@ days_with_snow = PrecipWithIndexing(
     description="{freq} number of days with snowfall larger than {low} and smaller or equal to {high}.",
     abstract="Number of days with snow between a lower and upper limit.",
     units="days",
-    compute=indices.days_with_snow,
+    compute=compute.days_with_snow,
 )
 
 snowfall_frequency = PrecipWithIndexing(
@@ -537,7 +537,7 @@ snowfall_frequency = PrecipWithIndexing(
     abstract="Percentage of days with snowfall above a given threshold (either a "
     "snowfall flux or a liquid water equivalent snowfall rate).",
     units="%",
-    compute=indices.snowfall_frequency,
+    compute=compute.snowfall_frequency,
 )
 
 snowfall_intensity = PrecipWithIndexing(
@@ -548,7 +548,7 @@ snowfall_intensity = PrecipWithIndexing(
     abstract="Mean daily liquid water equivalent snowfall rate above threshold (either a "
     "snowfall flux or a liquid water equivalent snowfall rate)",
     units="mm/day",
-    compute=indices.snowfall_intensity,
+    compute=compute.snowfall_intensity,
 )
 
 # FIXME: Are days_over_precip_thresh and days_over_precip_doy_thresh the same thing?
@@ -563,7 +563,7 @@ days_over_precip_thresh = PrecipWithIndexing(
     "calculated over a given period and a fixed threshold.",
     units="days",
     cell_methods="time: sum over days",
-    compute=indices.days_over_precip_thresh,
+    compute=compute.days_over_precip_thresh,
 )
 
 # FIXME: Are days_over_precip_thresh and days_over_precip_doy_thresh the same thing?
@@ -578,7 +578,7 @@ days_over_precip_doy_thresh = PrecipWithIndexing(
     abstract="Number of days in a period where precipitation is above a given daily percentile and a fixed threshold.",
     units="days",
     cell_methods="time: sum over days",
-    compute=indices.days_over_precip_thresh,
+    compute=compute.days_over_precip_thresh,
 )
 
 high_precip_low_temp = PrTasxWithIndexing(
@@ -589,7 +589,7 @@ high_precip_low_temp = PrTasxWithIndexing(
     abstract="Number of days with precipitation above a given threshold and temperature below a given threshold.",
     units="days",
     cell_methods="time: sum over days",
-    compute=indices.high_precip_low_temp,
+    compute=compute.high_precip_low_temp,
 )
 
 # FIXME: Are fraction_over_precip_thresh and fraction_over_precip_doy_thresh the same thing?
@@ -603,7 +603,7 @@ fraction_over_precip_doy_thresh = PrecipWithIndexing(
     "on each calendar day in the {pr_per_period} period, is used to compute the {pr_per_thresh}th percentile(s).",
     units="",
     cell_methods="",
-    compute=indices.fraction_over_precip_thresh,
+    compute=compute.fraction_over_precip_thresh,
 )
 
 # FIXME: Are fraction_over_precip_thresh and fraction_over_precip_doy_thresh the same thing?
@@ -615,7 +615,7 @@ fraction_over_precip_thresh = PrecipWithIndexing(
     "percentile of {pr_per_period} period. Only days with at least {thresh} are included in the total.",
     units="",
     cell_methods="",
-    compute=indices.fraction_over_precip_thresh,
+    compute=compute.fraction_over_precip_thresh,
 )
 
 liquid_precip_ratio = PrTasxWithIndexing(
@@ -627,7 +627,7 @@ liquid_precip_ratio = PrTasxWithIndexing(
     abstract="The ratio of total liquid precipitation over the total precipitation. Liquid precipitation is "
     "approximated from total precipitation on days where temperature is above a given threshold.",
     units="",
-    compute=indices.liquid_precip_ratio,
+    compute=compute.liquid_precip_ratio,
     parameters={"tas": {"kind": InputKind.VARIABLE}, "prsn": None},
 )
 
@@ -643,7 +643,7 @@ dry_spell_frequency = Precip(
     "over a given time window of days is below a given threshold.",
     units="",
     cell_methods="",
-    compute=indices.dry_spell_frequency,
+    compute=compute.dry_spell_frequency,
 )
 
 
@@ -658,7 +658,7 @@ dry_spell_total_length = Precip(
     "precipitation over a given time window of days is below a given threshold.",
     units="days",
     cell_methods="",
-    compute=indices.dry_spell_total_length,
+    compute=compute.dry_spell_total_length,
 )
 
 dry_spell_max_length = Precip(
@@ -672,7 +672,7 @@ dry_spell_max_length = Precip(
     "precipitation over a given time window of days is below a given threshold.",
     units="days",
     cell_methods="",
-    compute=indices.dry_spell_max_length,
+    compute=compute.dry_spell_max_length,
 )
 
 wet_spell_frequency = Precip(
@@ -686,7 +686,7 @@ wet_spell_frequency = Precip(
     "over a given time window of days is equal or above a given threshold.",
     units="",
     cell_methods="",
-    compute=indices.wet_spell_frequency,
+    compute=compute.wet_spell_frequency,
 )
 
 
@@ -701,7 +701,7 @@ wet_spell_total_length = Precip(
     "precipitation over a given time window of days is equal or above a given threshold.",
     units="days",
     cell_methods="",
-    compute=indices.wet_spell_total_length,
+    compute=compute.wet_spell_total_length,
 )
 
 wet_spell_max_length = Precip(
@@ -715,7 +715,7 @@ wet_spell_max_length = Precip(
     "precipitation over a given time window of days is equal or above a given threshold.",
     units="days",
     cell_methods="",
-    compute=indices.wet_spell_max_length,
+    compute=compute.wet_spell_max_length,
 )
 
 rprctot = PrecipWithIndexing(
@@ -729,7 +729,7 @@ rprctot = PrecipWithIndexing(
     "Only days with surpassing a minimum precipitation flux are considered.",
     units="",
     cell_methods="time: sum",
-    compute=indices.rprctot,
+    compute=compute.rprctot,
 )
 
 
@@ -743,7 +743,7 @@ cold_and_dry_days = PrecipWithIndexing(
     "precipitation is below {pr_per_thresh}th percentile.",
     abstract="Number of days with temperature below a given percentile and precipitation below a given percentile.",
     cell_methods="time: sum over days",
-    compute=indices.cold_and_dry_days,
+    compute=compute.cold_and_dry_days,
 )
 
 warm_and_dry_days = PrecipWithIndexing(
@@ -756,7 +756,7 @@ warm_and_dry_days = PrecipWithIndexing(
     "precipitation is below {pr_per_thresh}th percentile.",
     abstract="Number of days with temperature above a given percentile and precipitation below a given percentile.",
     cell_methods="time: sum over days",
-    compute=indices.warm_and_dry_days,
+    compute=compute.warm_and_dry_days,
 )
 
 warm_and_wet_days = PrecipWithIndexing(
@@ -769,7 +769,7 @@ warm_and_wet_days = PrecipWithIndexing(
     "precipitation is above {pr_per_thresh}th percentile.",
     abstract="Number of days with temperature above a given percentile and precipitation above a given percentile.",
     cell_methods="time: sum over days",
-    compute=indices.warm_and_wet_days,
+    compute=compute.warm_and_wet_days,
 )
 
 cold_and_wet_days = PrecipWithIndexing(
@@ -782,7 +782,7 @@ cold_and_wet_days = PrecipWithIndexing(
     "precipitation is above {pr_per_thresh}th percentile.",
     abstract="Number of days with temperature below a given percentile and precipitation above a given percentile.",
     cell_methods="time: sum over days",
-    compute=indices.cold_and_wet_days,
+    compute=compute.cold_and_wet_days,
 )
 
 rain_season = Precip(
@@ -810,7 +810,7 @@ rain_season = Precip(
     "The rain season starts with a period of abundant rainfall, followed by a period without prolonged dry sequences, "
     "which must happen before a given date. The rain season stops during a dry period happening after a given date.",
     cell_methods="",
-    compute=indices.rain_season,
+    compute=compute.rain_season,
 )
 
 water_cycle_intensity = PrecipWithIndexing(
@@ -822,7 +822,7 @@ water_cycle_intensity = PrecipWithIndexing(
     description="The {freq} water cycle intensity, defined as the sum of precipitation and actual evapotranspiration.",
     abstract="The sum of precipitation and actual evapotranspiration.",
     cell_methods="time: sum over days",
-    compute=indices.water_cycle_intensity,
+    compute=compute.water_cycle_intensity,
 )
 
 antecedent_precipitation_index = DailyPrecipNoResample(
@@ -833,5 +833,5 @@ antecedent_precipitation_index = DailyPrecipNoResample(
     description="Weighted moving sum of daily precipitation totals with a {window}-day window. Weights are an "
     "exponential decay of base {p_exp}.",
     cell_methods="time: sum over days",
-    compute=indices.antecedent_precipitation_index,
+    compute=compute.antecedent_precipitation_index,
 )