+++ src/xclim/indicators/land/_snow.py
@@ -2,7 +2,7 @@
 
 from __future__ import annotations
 
-from xclim import indices as xci
+from xclim import compute as xci
 from xclim.core.indicator import Daily, ResamplingIndicatorWithIndexing
 
 __all__ = [